changes prompted by the renaming of webgl/shader_funcs.ts to webgl/helper.ts
